automatic channel numbering (for applications such as VDR) for those services that support it.
Retrieves the channel numbers for the UK terrestrial "Freeview" service for use
in VDR 1.2.xx. The channel numbers can be found in the NIT table as the
user-defined descriptor 0x83 and each channel number has a corresponding
Service ID which is included in the descriptor.

patch contributed by John Murdoch